requirements.txt
# AgroVision AI — Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config
python-dotenv>=1.0.0

# AI
google-generativeai>=0.8.0
openai>=1.3.0

# Language
langdetect>=1.0.9

# Geo
geopy>=2.4.0

# ML
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
